+++ include/quadTree.h
@@ -238,7 +238,7 @@ public:
 	~QuadTreePowerLaw();
 
 protected:
-	float beta; // logorithmic slop of 2d mass profile
+	double beta; // logorithmic slop of 2d mass profile
 
 	// Override internal structure of halos
 	inline double alpha_h(double x,double xmax){
